Try a kind of factory pattern for preparing workflows.
