+++ gwf/core.py
@@ -4,8 +4,12 @@ import imp
 import inspect
 import os.path
 import sys
+from collections import defaultdict
 
-from .exceptions import TargetExistsError
+from .exceptions import (CircularDependencyError,
+                         FileProvidedByMultipleTargetsError,
+                         FileRequiredButNotProvidedError, TargetExistsError)
+from .utils import iter_inputs, iter_outputs
 
 _target_repr = (
     '{}(name={!r}, inputs={!r}, outputs={!r}, options={!r}, working_dir={!r}, '
@@ -41,6 +45,35 @@ def _norm_path(working_dir, path):
 
 def _norm_paths(working_dir, paths):
     return [_norm_path(working_dir, path) for path in paths]
+
+
+def _get_deep_dependencies(target, dependencies):
+    """Return all `Target`s necessary for building `target`.
+
+    The set of tasks is just returned as set.
+    """
+    # Working with a list to preserve the order. It makes lookups slower
+    # but hopefully these sets won't be terribly long ... if it becomes a
+    # problem it is easy enough to fix it.
+    processed = []
+
+    def dfs(other_target):
+        if other_target in processed:
+            return
+        else:
+            processed.append(other_target)
+            for dep in dependencies[other_target]:
+                dfs(dep)
+
+    dfs(target)
+    return processed
+
+
+def _check_circular_dependencies(workflow, dependencies):
+    for target in workflow.targets.values():
+        for dep in dependencies[target]:
+            if target in _get_deep_dependencies(dep, dependencies):
+                raise CircularDependencyError(target)
 
 
 class Target(object):
@@ -144,3 +177,41 @@ class Workflow(object):
         else:
             raise TypeError('First argument must be either a string or a '
                             'Workflow object.')
+
+
+def prepare_workflow(workflow):
+    provides = {}
+    dependencies = defaultdict(list)
+    dependents = defaultdict(list)
+
+    for target, path in iter_outputs(workflow.targets.values()):
+        if path in provides:
+            raise FileProvidedByMultipleTargetsError(
+                path, provides[path].name, target
+            )
+
+        provides[path] = target
+
+    for target, path in iter_inputs(workflow.targets.values()):
+        if os.path.exists(path):
+            continue
+
+        if path not in provides:
+            raise FileRequiredButNotProvidedError(path, target)
+        dependencies[target].append(provides[path])
+
+    for target, deps in dependencies.items():
+        for dep in deps:
+            dependents[dep].append(target)
+
+    for target in workflow.targets.values():
+        for dep in dependencies[target]:
+            if target in _get_deep_dependencies(dep, dependencies):
+                raise CircularDependencyError(target)
+
+    # Attach prepared attributes to the workflow.
+    setattr(workflow, 'provides', provides)
+    setattr(workflow, 'dependencies', dependencies)
+    setattr(workflow, 'dependents', dependents)
+
+    return workflow
